+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Version 1.2.0-dev
 - Added PQC signing: supported for now are ML-DSA ([RFC 9882](https://www.rfc-editor.org/info/rfc9882/)) and SLH-DSA ([RFC 9814](https://www.rfc-editor.org/info/rfc9814/))
+- Added `PdfSigningAlgorithm`, deprecating `PdfSignatureEncryption`
 
 ## Version 1.1.1
 - `PdfColorSpaceFilterIndexed`: Added support for fetching `/DeviceRGB` images with `/BitsPerComponent` != 8